ST-NS-014 Fix remaining mypy errors in codebase
- Fix pre_market_briefing.py: add explicit float() cast
- Fix confidence_filter.py: add explicit bool() cast
- Fix signal_history/tracker.py: add type: ignore for storage returns
- Fix markov/probability_calculator.py: add explicit float() cast
- Fix data_validator.py: add explicit float() cast

+++ src/market_analysis/markov/probability_calculator.py
@@ -406,7 +406,7 @@ class ProbabilityCalculator:
 
         if prob > 0:
             # Geometric distribution expected value
-            return 1.0 / prob
+            return float(1.0 / prob)
 
         # If direct transition probability is 0, estimate from matrix structure
         # This is a simplified approximation
